Create PrimeChecker
This Java program consists of two classes: Solution and Prime.

Class Solution: This class contains the main method which is the entry point for the program.

A Scanner object sc is created to read input from the user.
Five integers n1, n2, n3, n4, and n5 are read from the input.
The static method checkPrime of the Prime class is called four times with different numbers of arguments.
Class Prime: This class contains two static methods: checkPrime and isPrime.

checkPrime(int... numbers): This method takes a variable number of integer arguments. It iterates over each number and checks if it’s prime by calling the isPrime method. If the number is prime, it’s printed to the console. After checking all numbers, it prints a newline.

isPrime(int n): This method checks if a given number n is prime. It counts the number of divisors of n. If n has exactly two divisors (which means it’s prime), the method returns true. Otherwise, it returns false.